% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_MSdial_to_Jupyter_app}
\alias{run_MSdial_to_Jupyter_app}
\title{Run the MSdial To Jupyter Application}
\usage{
run_MSdial_to_Jupyter_app()
}
\value{
A Shiny application object
}
\description{
This function launches the Shiny application for converting MSdial output to Jupyter format.
The application allows users to upload MSdial CSV output and convert it to a format
compatible with Jupyter notebooks.
}
\examples{
if(interactive()){
  run_MSdial_to_Jupyter_app()
}
}
